requirements.txt
torch
torchvision
transformers==4.44.2
lancedb
sentence-transformers
opencv-python
pillow